requirements.txt
torch
torchvision
pandas 
opencv-python
tqdm
matplotlib
pyqt5
pillow
scipy
scikit-learn
